requirements.txt
absl-py==2.1.0
aiohttp==3.9.2
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
ase==3.22.1
async-timeout==4.0.3
attrs==23.2.0
backports.cached-property==1.0.2
black==24.1.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dill==0.3.8
docker-pycreds==0.4.0
element-coder==0.0.8
f90wrap==0.2.13
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
gmpy2==2.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.60.0
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lammps==2023.11.21
lammps-interface==0.2.2
latexcodec==2.0.1
libconeangle==0.1.2
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.7.2
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
mendeleev==0.15.0
mofchecker==0.9.6
mofdiff==1.0.0
monty==2024.1.26
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
p_tqdm==1.4.0
packaging==23.2
palettable==3.3.3
pandas==2.2.0
pathos==0.3.2
pathspec==0.12.1
pillow==10.2.0
pip==23.3.2
platformdirs==4.1.0
plotly==5.18.0
ply==3.11
pox==0.3.4
ppft==1.7.6.8
protobuf==4.24.4
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pybtex==0.24.0
pycparser==2.21
pyDeprecate==0.3.2
pyeqeq==0.0.10
pyfiglet==0.8.post1
Pygments==2.17.2
PyJWT==2.8.0
pymatgen==2024.1.27
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==1.6.1
pytz==2023.4
pyu2f==0.1.5
PyYAML==6.0.1
quippy-ase==0.9+gitnone
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.7
scikit-learn==1.4.0
scipy==1.12.0
sentry-sdk==1.40.0
setproctitle==1.3.3
setuptools==59.5.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
spglib==2.3.0
SQLAlchemy==2.0.25
structuregraph-helpers==0.0.9
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.0
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torch_geometric==2.4.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torchmetrics==0.7.3
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uncertainties==3.1.7
unicodedata2==15.1.0
urllib3==2.2.0
wandb==0.16.2
Werkzeug==3.0.1
wheel==0.42.0
yarl==1.9.4
zipp==3.17.0
